refactor: improve import and export logic
